+++ example/lib/page/webview_page.dart
@@ -8,6 +8,6 @@ class FlWebViewPage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return BasicScaffold(appBarTitle: 'FlWebView', isScroll: isCalculateHeight);
+    return BasicScaffold(appBarTitleText: 'FlWebView', isScroll: isCalculateHeight);
   }
 }